Create README.md
Feel free to use it as is; this REST API is fully functional and designed to generate an image with a YouTube link as a parameter. Simply set the slide duration in seconds to capture the desired image. For any questions or additional details, don't hesitate to reach out to me at admin@wimsup.com or visit https://wimsup.com. Thank you, and best of luck!

Signed-off-by: Milind Bhave <mcbhave@hotmail.com>